refactor: added folders for mappings in application layer
