+++ src/Application/Mappings/Initiatives/InitiativeUserMappings.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Mappings;
+﻿namespace IAVH.BioTablero.CM.Application.Mappings.Initiatives;
 
 using System;
 
@@ -7,28 +7,28 @@ using IAVH.BioTablero.CM.Application.DTOs.Utils;
 using IAVH.BioTablero.CM.Application.Interfaces.General;
 using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
 
-using TagCategoryEnum = IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums.TagCategory;
+using InitiativeUserLevelEnum = IAVH.BioTablero.CM.Core.Domain.Utils.Enums.InitiativesEnums.InitiativeUserLevel;
 
 /// <summary>
-/// Tag mappings.
+/// Initiative contact mappings.
 /// </summary>
-public class TagMappings : IMapper<Tag, TagDto>
+public class InitiativeUserMappings : IMapper<InitiativeUser, InitiativeUserDto>
 {
     /// <summary>
     /// Map from entity to DTO.
     /// </summary>
     /// <param name="entity">Entity data.</param>
     /// <returns>DTO data.</returns>
-    public TagDto Map(Tag entity)
+    public InitiativeUserDto Map(InitiativeUser entity)
     {
         ArgumentNullException.ThrowIfNull(entity);
 
         return new()
         {
             Id = entity.Id,
-            Name = entity.Name,
-            Url = entity.Url.ToString(),
-            Category = new EnumEntityDto<TagCategoryEnum>(entity.CategoryId),
+            InitiativeId = entity.InitiativeId,
+            UserName = entity.UserName,
+            Level = new EnumEntityDto<InitiativeUserLevelEnum>(entity.LevelId),
         };
     }
 
@@ -37,15 +37,15 @@ public class TagMappings : IMapper<Tag, TagDto>
     /// </summary>
     /// <param name="dto">DTO data.</param>
     /// <returns>Entity data.</returns>
-    public Tag Map(TagDto dto)
+    public InitiativeUser Map(InitiativeUserDto dto)
     {
         ArgumentNullException.ThrowIfNull(dto);
 
         return new()
         {
-            Name = dto.Name,
-            Url = new Uri(dto.Url),
-            CategoryId = dto.Category.Id,
+            InitiativeId = dto?.InitiativeId ?? 0,
+            UserName = dto.UserName,
+            LevelId = dto.Level.Id,
         };
     }
 }
