+++ src/Application/Mappings/TerritoryStory/TerritoryStoryVideoMappings.cs
@@ -1,4 +1,4 @@
-﻿namespace IAVH.BioTablero.CM.Application.Mappings;
+﻿namespace IAVH.BioTablero.CM.Application.Mappings.TerritoryStory;
 
 using System;
