+++ src/WebApi/Config/DependencyRegistry/ConfigMappings.cs
@@ -5,7 +5,10 @@ using IAVH.BioTablero.CM.Application.DTOs.Initiatives;
 using IAVH.BioTablero.CM.Application.DTOs.Logging;
 using IAVH.BioTablero.CM.Application.DTOs.TerritoryStories;
 using IAVH.BioTablero.CM.Application.Interfaces.General;
-using IAVH.BioTablero.CM.Application.Mappings;
+using IAVH.BioTablero.CM.Application.Mappings.Geo;
+using IAVH.BioTablero.CM.Application.Mappings.Initiatives;
+using IAVH.BioTablero.CM.Application.Mappings.Logging;
+using IAVH.BioTablero.CM.Application.Mappings.TerritoryStory;
 using IAVH.BioTablero.CM.Core.Domain.Entities.Geo;
 using IAVH.BioTablero.CM.Core.Domain.Entities.Initiatives;
 using IAVH.BioTablero.CM.Core.Domain.Entities.Logging;
